Break out tessellation control to its own file for reuse
