+++ CMakeLists.txt
@@ -878,6 +878,7 @@ set(CORE_SOURCES
   src/core/SourceFileCache.cc
   src/core/StatCache.cc
   src/core/SurfaceNode.cc
+  src/core/TessellationControl.cc
   src/core/TextNode.cc
   src/core/TransformNode.cc
   src/core/Tree.cc
