+++ src/core/primitives.cc
@@ -27,7 +27,6 @@
 #include "core/primitives.h"
 
 #include "geometry/Geometry.h"
-#include "geometry/Grid.h"
 #include "geometry/linalg.h"
 #include "geometry/PolySet.h"
 #include "geometry/Polygon2d.h"
@@ -53,8 +52,6 @@
 #include <vector>
 
 using namespace boost::assign;  // bring 'operator+=()' into scope
-
-#define F_MINIMUM 0.01
 
 template <class InsertIterator>
 static void generate_circle(InsertIterator iter, double r, double z, int fragments)
@@ -174,7 +171,7 @@ std::unique_ptr<const Geometry> SphereNode::createGeometry() const
     return PolySet::createEmpty();
   }
 
-  int num_fragments = fragments.subdivisions_from_r(r).value_or(3);
+  int num_fragments = fragments.circular_segments(r).value_or(3);
   auto num_rings = (num_fragments + 1) / 2;
   // Uncomment the following three lines to enable experimental sphere
   // tessellation
@@ -220,7 +217,7 @@ static std::shared_ptr<AbstractNode> builtin_sphere(const ModuleInstantiation *i
 {
   Parameters parameters = Parameters::parse(std::move(arguments), inst->location(), {"r"}, {"d"});
 
-  auto node = std::make_shared<SphereNode>(inst, Fragments(parameters, inst));
+  auto node = std::make_shared<SphereNode>(inst, TessellationControl(parameters, inst));
 
   const auto r = lookup_radius(parameters, inst, "d", "r");
   if (r.type() == Value::Type::NUMBER) {
@@ -241,7 +238,7 @@ std::unique_ptr<const Geometry> CylinderNode::createGeometry() const
     return PolySet::createEmpty();
   }
 
-  int num_fragments = fragments.subdivisions_from_r(std::fmax(this->r1, this->r2)).value_or(3);
+  int num_fragments = fragments.circular_segments(std::fmax(this->r1, this->r2)).value_or(3);
 
   double z1, z2;
   if (this->center) {
@@ -298,7 +295,7 @@ static std::shared_ptr<AbstractNode> builtin_cylinder(const ModuleInstantiation 
 {
   Parameters parameters = Parameters::parse(std::move(arguments), inst->location(),
                                             {"h", "r1", "r2", "center"}, {"r", "d", "d1", "d2"});
-  auto node = std::make_shared<CylinderNode>(inst, Fragments(parameters, inst));
+  auto node = std::make_shared<CylinderNode>(inst, TessellationControl(parameters, inst));
 
   if (parameters["h"].type() == Value::Type::NUMBER) {
     node->h = parameters["h"].toDouble();
@@ -542,7 +539,7 @@ std::unique_ptr<const Geometry> CircleNode::createGeometry() const
     return std::make_unique<Polygon2d>();
   }
 
-  int num_fragments = fragments.subdivisions_from_r(this->r).value_or(3);
+  int num_fragments = fragments.circular_segments(this->r).value_or(3);
   Outline2d o;
   o.vertices.resize(num_fragments);
   for (int i = 0; i < num_fragments; ++i) {
@@ -555,7 +552,7 @@ std::unique_ptr<const Geometry> CircleNode::createGeometry() const
 static std::shared_ptr<AbstractNode> builtin_circle(const ModuleInstantiation *inst, Arguments arguments)
 {
   Parameters parameters = Parameters::parse(std::move(arguments), inst->location(), {"r"}, {"d"});
-  auto node = std::make_shared<CircleNode>(inst, Fragments(parameters, inst));
+  auto node = std::make_shared<CircleNode>(inst, TessellationControl(parameters, inst));
 
   const auto r = lookup_radius(parameters, inst, "d", "r");
   if (r.type() == Value::Type::NUMBER) {
@@ -760,45 +757,4 @@ void register_builtin_primitives()
                    "polygon([points])",
                    "polygon([points], [paths])",
                  });
-}
-
-Fragments::Fragments(const Parameters& parameters, const ModuleInstantiation *inst)
-{
-  fn = parameters["$fn"].toDouble();
-  fs = parameters["$fs"].toDouble();
-  fa = parameters["$fa"].toDouble();
-
-  if (fs < F_MINIMUM) {
-    if (inst) {
-      LOG(message_group::Warning, inst->location(), parameters.documentRoot(),
-          "$fs too small - clamping to %1$f", F_MINIMUM);
-    }
-    fs = F_MINIMUM;
-  }
-  if (fa < F_MINIMUM) {
-    if (inst) {
-      LOG(message_group::Warning, inst->location(), parameters.documentRoot(),
-          "$fa too small - clamping to %1$f", F_MINIMUM);
-    }
-    fa = F_MINIMUM;
-  }
-}
-
-/*!
-   Returns the number of subdivision of a whole circle, given radius and
-   the three special variables $fn, $fs and $fa
- */
-std::optional<int> Fragments::subdivisions_from_r(double r) const
-{
-  // FIXME: It would be better to refuse to create an object. Let's do more strict error handling
-  // in future versions of OpenSCAD
-  if (r < GRID_FINE || std::isinf(fn) || std::isnan(fn)) return {};
-  if (fn > 0.0) return static_cast<int>(fn >= 3 ? fn : 3);
-  return static_cast<int>(ceil(fmax(fmin(360.0 / fa, r * 2 * M_PI / fs), 5)));
-}
-
-std::ostream& operator<<(std::ostream& stream, const Fragments& f)
-{
-  stream << "$fn = " << f.fn << ", $fa = " << f.fa << ", $fs = " << f.fs;
-  return stream;
 }
